requirements.txt
# Core dependencies
botocore==1.34.162 
boto3==1.34.162 
ipywidgets==8.1.5
python-dotenv==1.0.1
IPython>=8.12.0

# Image processing
pillow==11.1.0
pdf2image==1.17.0

# Machine Learning & AI
torch
transformers
huggingface-hub
# Vector Database
qdrant-client

# Visualization
matplotlib
tqdm==4.67.1


# Optional: Local LLM support
ollama

# Additional dependencies for enhanced RAG and agent functionality
langchain
langchain-aws
duckduckgo_search
scikit-learn==1.6.1
pandas==2.2.3
numpy
sentence-transformers==3.4.1

PyPDF2==3.0.1

# Colpali
colpali-engine
colpali-engine[interpretability]